+++ components/Login/Login.js
@@ -1,30 +1,13 @@
 import { useState } from "react";
-import { GoogleAuthProvider, getAuth, signInAnonymously, signInWithPopup, updateProfile } from "firebase/auth";
 
+import { useAuth } from "@/providers/AuthProvider";
 import Button from "../Button/Button";
 import Input from "../Input/Input";
 import styles from "./Login.module.css";
 
 const Login = () => {
-  const auth = getAuth();
-
-  const anonymousSignIn = async (event) => {
-    try {
-      if (displayName) {
-        await signInAnonymously(auth);
-        await updateProfile(auth.currentUser, { displayName });
-      }
-    } catch (e) {}
-  };
-
-  const googleSignIn = async (event) => {
-    try {
-      const provider = new GoogleAuthProvider();
-      signInWithPopup(auth, provider);
-    } catch (e) {}
-  };
-
   const [displayName, setDisplayName] = useState("");
+  const { anonymousSignIn, googleSignIn } = useAuth();
 
   return (
     <div className={styles.root}>
@@ -33,13 +16,13 @@ const Login = () => {
         <label>Display Name</label>
         <Input type="text" value={displayName} onChange={(e) => setDisplayName(e.target.value)}></Input>
       </div>
-      <Button className={styles.anonButton} onClick={anonymousSignIn} disabled={!displayName}>
+      <Button className={styles.anonButton} onClick={() => anonymousSignIn(displayName)} disabled={!displayName}>
         Anonymous Sign In
       </Button>
       <div>
-        <span> - - - - - - - - - - - - - - - - - - -</span>
+        <span> - - - - - - - - - - - - - - - - - - - - - </span>
         OR
-        <span> - - - - - - - - - - - - - - - - - - -</span>
+        <span> - - - - - - - - - - - - - - - - - - - - </span>
       </div>
       <div className={styles.other}>
         <Button onClick={googleSignIn}>Continue with Google</Button>
